refactor(Application): 파일 분리
